Google Earth Version/requirements.txt
absl-py==1.3.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
Babel==2.10.3
babelgrid==0.1.5
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==4.0.2
bleach==5.0.1
blinker==1.4
branca==0.5.0
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.4.0

earthengine-api==0.1.327
entrypoints==0.3
executing==1.1.0
fastjsonschema==2.16.2

flatbuffers==22.11.23
folium==0.13.0
fonttools==4.38.0
future==0.18.2
gast==0.4.0
geetools==0.6.14
geojson==2.5.0
geopandas==0.10.2
google-api-core==2.10.2
google-api-python-client==1.12.11
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
googlemaps==4.7.3
grpcio==1.50.0
h3==3.6.3
h5py==3.7.0
html5lib==1.0.1
httplib2==0.20.4
httplib2shim==0.0.3
hyperlink==19.0.0
idna==2.8
importlib-metadata==5.0.0
incremental==16.10.1
ipykernel==6.16.0
ipyleaflet==0.17.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-console==6.0.0
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.2
jupyterthemes==0.20.0
keras==2.11.0
keyring==18.0.1
kiwisolver==1.4.4

launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lesscpy==0.15.1
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==4.2.0
munch==2.5.0
nbclassic==0.4.4
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
opencv-contrib-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pbr==5.10.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pyGeoTile==1.0.6
Pygments==2.13.0

PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9

pyrsistent==0.15.5
pyserial==3.4


python-dateutil==2.8.2

pytz==2022.4
PyYAML==5.3.1
pyzmq==24.0.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
Rtree==1.0.1
s2sphere==0.2.5
scikit-learn==1.1.3
scipy==1.9.3
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
Shapely==1.8.5.post1
simplejson==3.16.0
six==1.14.0
scikit-learn
sniffio==1.3.0
sos
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.5.1
svgwrite==1.4.3
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.16.0
testpath==0.4.4
testresources==2.0.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
traittypes==0.2.1
Tree==0.2.4
Twisted==18.9.0
typing_extensions==4.4.0

unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xyzservices==2022.9.0
zipp==1.0.0
